pyproject.toml
# ==== Project metadata ====

[project]
name = "validibot-validators"
version = "0.5.0"
description = "Validibot advanced validator containers (EnergyPlus, FMU)"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Daniel McQuillen", email = "daniel@validibot.com" },
    { name = "McQuillen Interactive Pty. Ltd." },
]
keywords = [
    "validibot",
    "validation",
    "energyplus",
    "fmi",
    "fmu",
    "simulation",
    "building-energy",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Quality Assurance",
]

# Core dependencies shared by all validators
dependencies = [
    "pydantic==2.12.3",
    "httpx==0.28.1",
    "google-cloud-storage==3.9.0",
    "google-auth==2.48.0",
    "requests==2.32.5",
    "sentry-sdk==2.52.0",
    "validibot-shared==0.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-mock==3.15.1",
    "pytest-cov==7.0.0",
    "ruff==0.15.0",
    "mypy==1.19.1",
]

# EnergyPlus validator dependencies (for local dev/testing)
energyplus = [
    # EnergyPlus-specific deps go here
]

# FMU validator dependencies (for local dev/testing)
fmu = ["fmpy==0.3.27"]

[project.urls]
Homepage = "https://validibot.com"
Documentation = "https://docs.validibot.com"
Repository = "https://github.com/danielmcquillen/validibot-validators"
Issues = "https://github.com/danielmcquillen/validibot-validators/issues"

[tool.uv.sources]
# validibot-shared: Prefer the sibling checkout during coordinated local
# development so validators are tested against the current shared contract.
validibot-shared = { path = "../validibot-shared", editable = true }

# ==== Build system ====

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["validators"]

# ==== pytest ====

[tool.pytest.ini_options]
testpaths = [
    "tests",
    "validators/energyplus/tests",
    "validators/core/tests",
    "validators/fmu/tests",
]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (may require Docker/containers)",
    "slow: Slow tests (skipped by default)",
]
filterwarnings = ["ignore::DeprecationWarning"]

# ==== Coverage ====

[tool.coverage.run]
source = ["validators"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ==== Ruff ====

[tool.ruff]
target-version = "py313"
line-length = 99
src = ["validators"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate (commented-out code)
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults
    "ARG001", # Unused function argument (common in interfaces)
    "ARG002", # Unused method argument
]

[tool.ruff.lint.isort]
known-first-party = ["validators", "validibot_shared"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",  # Unused arguments OK in tests (fixtures)
    "S101", # Assert OK in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# ==== mypy ====

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = false
exclude = ["tests/", "build/", "dist/"]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
